Make trainee maintainers three times as likely to be picked
Trainee maintainers count as reviewers, but should get more reviews than
most reviewers, as they need practice to become a maintainer. This makes
them three times as likely to be picked (compared to another reviewer;
compared to before, it's roughly twice as likely at the current numbers
for each).

Also, switch to `Array#sample` because I think it looks nicer.