Don't fail the job if the Review App doesn't deploy properly
It seems the deploy function causes the job to fail if it doesn't
succeed. That wasn't the intent as we want to curl the Review App after
the deploy finished (even if it failed) because sometimes the Review App
is just a bit long to be ready.

This change wraps the Review App deployment with "set +e"/"set -e" to
ensure that the job doesn't fail right away if the deploy fails.

Signed-off-by: Rémy Coutable <remy@rymai.me>