Remove unnecessary comments from login tests
The comments were TODOs but there were no explanation of what
needed to be done.

Also, I think the expectation describes very well how it checks
that the user is logged in.

Finally, as Robert C. Martin said once:
"After you check in a TODO comment it turns into a DON'T DO"