Include all queries in the performance bar summary
Rails 5.1's `exec_no_cache` doesn't call `async_exec` any more, but
`exec_params`: https://github.com/rails/rails/pull/33188

This means that the DB summary in the performance bar was wrong. The
individual query details were still correct (we subscribe to
ActiveRecord events for those).

We can remove this once the upstream PR to peek-pg is in a release and
we update to use that release.