Additional fix to handle NULL lock_version
If the UI sends a string value for lock_version (e.g. "0"), then the
previous monkey patch did not properly handle that properly. This
commit casts the value to an integer to determine whether to look for
NULL lock_versions.

For merge requests, GitLab sends a POST request to
`namespace/project/merge_requests/:iid` with the
`merge_request[lock_version]` parameter with a string `0`. The string
value comes from the form field, which explains why
https://gitlab.com/gitlab-org/gitlab-ce/merge_requests/28145 wasn't
sufficient.