Whitelist Gitaly N+1 for forking in tests
This helper is right on the limit of the number of calls being made.
When the request store is enabled, and Sidekiq is running in inline
mode, just a couple of additional actions result in it being pushed
over the line.

The operation is entirely artificial, since in reality we'd perform
these requests in multiple separate processes, so just whitelist it
for now.