Add missing sprintf and remove bad i18n
An instance of i18n which required sprintf was not handled correctly
and has been fixed in this commit.

Further to this, i18n was applied to
  app/assets/javascripts/new_branch_form.js
but needed to be removed as the i18n was too vague.

A note has been made to follow up on this file.