Fix migration failure when groups are missing route
When a Group is missing a route, the migration failed previously with a
`Invalid single-table inheritance type` error. To fix this, we can
disable STI for the migration class because we don't need to know about
the type to do this migration. Besides, currently Group is the only type
used in the type column.

Closes https://gitlab.com/gitlab-org/gitlab-ce/issues/58714