Don't import pull request comments from Gitea repos
The Gitea API does not provide the following API endpoint for pull request
comments:

  /api/v1/repos/{owner}/{repo}/pulls/comments

When the importer attempts to request this endpoint it receives a '404
Not Found' error which causes the import to fail. By skipping any
attempts to import pull requests comments from Gitea we can ensure that
the import can complete successfully.