Leave clear notices about after_script execution context
We frequently see new issues mentioning that something defined
in `before_script` or `script` doesn't work with `after_script`.

Our documentation mentions that first two are concatenated and
executed together, why the third one is executed in a separated
shell context. But it looks like it's not clear enough.

Since this misunderstanding of how scripts are being executed
is repeated frequently, let's make it more clear in docs.