Update to node@12
This updates our build images to use node@12. We are also adding a
sentence to the from-source installation that while we require a minimum
version of node@8, we also recommend node@12. Last but not least, we
change the version of node in the .nvmrc so that developers using nvm
can automatically use node@12 with the project