Add a danger rule to suggest `docs-` prefixes
If an MR only changes files within `doc/` we can save valueable CI time
by prepending a branch name with `docs-`. This danger rule helps educate
people about that fact.

On the other hand, if someone has a branch with a `docs` prefix or
suffix, the branch shouldn't contain any changes outside of the `docs`
directory